Crate bp_bridge_hub_westend

Module with configuration which reflects BridgeHubWestend runtime setup (AccountId, Headers, Hashes…)

Structs

AccountInfoStorageMapKeyProvider

Provides a storage key for account data.

BlockExecutionWeight

Importing a block with 0 Extrinsics.

BlockLength

Size limit of the Cumulus-based bridge hub blocks.

BlockWeights

Weight limit of the Cumulus-based bridge hub blocks.

BlockWeightsForAsyncBacking

Weight limit of the Cumulus-based bridge hub blocks when async backing is enabled.

BridgeHubWestend

BridgeHubWestend parachain.

BridgeHubWestendBaseConfirmationFeeInWnds

Transaction fee that is paid at the Westend BridgeHub for delivering single outbound message confirmation. (initially was calculated by test BridgeHubWestend::can_calculate_fee_for_standalone_message_confirmation_transaction + 33%)

BridgeHubWestendBaseDeliveryFeeInWnds

Transaction fee that is paid at the Westend BridgeHub for delivering single inbound message. (initially was calculated by test BridgeHubWestend::can_calculate_fee_for_standalone_message_delivery_transaction + 33%)

BridgeHubWestendBaseXcmFeeInWnds

The XCM fee that is paid for executing XCM program (with ExportMessage instruction) at the Westend BridgeHub. (initially was calculated by test BridgeHubWestend::can_calculate_weight_for_paid_export_message_with_reserve_transfer + 33%)

ExtrinsicBaseWeight

Executing a NO-OP System::remarks Extrinsic.

Perbill

A fixed point representation of a number in the range [0, 1].

Enums

RuntimeCall

Wrapper over BridgeHubWestend’s RuntimeCall that can be used without a runtime.

Constants

AVERAGE_BLOCK_INTERVAL

Average block interval in Cumulus-based parachains.

AVERAGE_ON_INITIALIZE_RATIO

All cumulus bridge hubs assume that about 5 percent of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

BEST_FINALIZED_BRIDGE_HUB_WESTEND_HEADER_METHOD

Name of the <ThisChain>FinalityApi::best_finalized runtime method.

BRIDGE_HUB_WESTEND_PARACHAIN_ID

Identifier of BridgeHubWestend in the Westend relay chain.

EXTRA_STORAGE_PROOF_SIZE

Number of extra bytes (excluding size of storage value itself) of storage proof, built at Polkadot-like chain. This mostly depends on number of entries in the storage trie. Some reserve is reserved to account future chain growth.

FREE_HEADERS_INTERVAL_FOR_BRIDGE_HUB_WESTEND_METHOD

Name of the <ThisChain>FinalityApi::free_headers_interval runtime method.

FROM_BRIDGE_HUB_WESTEND_MESSAGE_DETAILS_METHOD

Name of the From<ThisChain>InboundLaneApi::message_details runtime method.

MAX_BRIDGE_HUB_HEADER_SIZE

Maximal bridge hub header size.

MAX_UNCONFIRMED_MESSAGES_IN_CONFIRMATION_TX

Maximal number of unconfirmed messages at inbound lane for Cumulus-based parachains. Note: this value is security-relevant, decreasing it should not be done without careful analysis (like the one above).

MAX_UNREWARDED_RELAYERS_IN_CONFIRMATION_TX

Maximal number of unrewarded relayer entries at inbound lane for Cumulus-based parachains. Note: this value is security-relevant, decreasing it should not be done without careful analysis (like the one above).

NORMAL_DISPATCH_RATIO

All cumulus bridge hubs allow normal extrinsics to fill block up to 75 percent.

TO_BRIDGE_HUB_WESTEND_MESSAGE_DETAILS_METHOD

Name of the To<ThisChain>OutboundLaneApi::message_details runtime method.

TX_EXTRA_BYTES

Maximal number of bytes, included in the signed Polkadot-like transaction apart from the encoded call itself.

WITH_BRIDGE_HUB_WESTEND_MESSAGES_PALLET_NAME

Name of the With-BridgeHubWestend messages pallet instance that is deployed at bridged chains.

WITH_BRIDGE_HUB_WESTEND_RELAYERS_PALLET_NAME

Name of the With-BridgeHubWestend bridge-relayers pallet instance that is deployed at bridged chains.

WITH_BRIDGE_WESTEND_TO_ROCOCO_MESSAGES_PALLET_INDEX

Pallet index of BridgeRococoMessages: pallet_bridge_messages::<Instance1>.

Traits

BridgeHubWestendFinalityApi

API for querying information about the finalized chain headers.

FromBridgeHubWestendInboundLaneApi

Inbound message lane API for messages sent by this chain.

ToBridgeHubWestendOutboundLaneApi

Outbound message lane API for messages that are sent to this chain.

Type Aliases

AccountId

Id of account on Polkadot-like chains.

AccountPublic

Public key of account on Polkadot-like chains.

AccountSigner

Public key of the chain account that may be used to verify signatures.

Address

The address format for describing accounts.

Balance

The balance of an account on Polkadot-like chain.

BlockNumber

Block number type used in Polkadot-like chains.

Hash

Hash type used in Polkadot-like chains.

Hasher

The type of object that can produce hashes on Polkadot-like chains.

Hashing

Hashing type.

Header

The header type used by Polkadot-like chains.

Nonce

Nonce of a transaction on the Polkadot-like chains.

Signature

Signature type used by Polkadot-like chains.

SignedBlock

Polkadot-like block signed with a Justification.

TransactionExtension

Signed extension that is used by all bridge hubs.

UncheckedExtrinsic

Unchecked Extrinsic type.